Jekyll to Hugo bash completed

+++ resources/migrate_jekyll_to_hugo.sh
@@ -158,14 +158,22 @@ sed -i.bak "s|^theme = .*$|theme = \"${HUGO_THEME_NAME}\"|" config.toml
 # grep -q '^option' file && sed -i 's/^option.*/option=value/' file || echo 'option=value' >> file
 grep -q '^publishDir' config.toml && sed -i "s|^publishDir = .*$|publishDir = \"../${HUGO_CONTENT_DIR}/docs\"|" config.toml || sed -i "s|^\(theme = .*\)$|\1\npublishDir = \"../${HUGO_CONTENT_DIR}/docs\"|" config.toml
 
-printf "\n"
-echo "---------------------------------------------------------------"
-echo " Main branch - Removing not used files (only for https://github.com/chilcano/ghpages-holosec.git)"
-echo "---------------------------------------------------------------"
-rm -rf static/assets/main.scss
-rm -rf static/assets/fonts/
-rm -rf static/wp_export/
-rm -rf static/CNAME
+
+if [[ "${GHREPO_SOURCE_JEKYLL_URL,,}" =~  .*"github.com/chilcano/ghpages-holosec".* ]]; then
+  printf "\n"
+  echo "---------------------------------------------------------------"
+  echo " Main branch - Removing unused files "
+  echo " (only for https://github.com/chilcano/ghpages-holosec.git)"
+  echo "---------------------------------------------------------------"
+  rm -rf static/wp_export/
+  rm -rf static/CNAME
+  rm -rf static/assets/fonts/
+  rm -rf static/assets/main.scss
+  find static/assets/ -maxdepth 1 -type f -name '*.png' -o -name '*.jpg' -o -name '*.pdf' -exec rm -rf {} \;
+  find static/assets/ -type d -name 'blog201*' -o -name 'blog' -exec rm -rf {} \;
+  find content/post -maxdepth 1 -type f -name '200[7-9]*' -exec rm -rf {} \;
+  find content/post -maxdepth 1 -type f -name '201[0-8]*' -exec rm -rf {} \;
+fi
 
 printf "==> Changing to '${PATH_TARGET_REPO}/' as working directory. \n"
 cd ${PATH_TARGET_REPO}/
